requirements.txt
beautifulsoup4
wikipedia
mysql-connector-python

flask
flask-restplus

Sphinx
autodocsumm